+++ packages/core/src/hid/index.ts
@@ -1,3 +1,4 @@
 export * from "./hid-buttons";
 export * from "./hid-gamepad";
+export * from "./playstation";
 export * from "./xinput";